requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
youtube-dl==2021.12.17
pytube==15.0.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
yt-dlp==2023.11.16
